test(jina): split provider tests
