+++ specs/GH727/product.md
@@ -11,33 +11,33 @@ GH-727 / #727
 
 ## 本 tranche 目标
 
-- 拆分 `src/core/providers/azure/batches/mod.rs`，它当前 1193 行，是 #727 当前 top offenders 之一。
-- 按现有 `#[path = "..._tests.rs"] mod tests;` 模式，把 inline tests 移动到 `batches_tests.rs`。
-- 保持 Azure Batch runtime 代码、测试断言、fixtures 和 public API 不变。
+- 拆分 `src/core/providers/jina/mod.rs`，它当前 1164 行，是 #727 当前 top offenders 之一。
+- 按现有 `#[path = "..."] mod tests;` 模式，把 inline tests 移动到 `tests.rs`。
+- 保持 Jina runtime 代码、测试断言、fixtures 和 public API 不变。
 - 所有新增或修改后的 Rust 文件低于 800 行。
 
 ## 非目标
 
-- 不修改 Azure Batch runtime 行为。
-- 不重构 Azure Batch API surface、request/response types、URL building logic 或 validation rules。
+- 不修改 Jina runtime 行为。
+- 不重构 Jina API surface、provider config、model catalog、rerank DTOs 或 error mapping。
 - 不在本 PR 中处理其余大文件。
 - 不关闭 #727，除非 issue owner 决定一个 tranche PR 足以满足 tracker。
 
 ## Behavior Invariants
 
-1. 所有原有 Azure Batch tests 仍由 `core::providers::azure::batches::tests::*` 测试树运行。
+1. 所有原有 Jina tests 仍由 `core::providers::jina::tests::*` 测试树运行。
 2. 测试移动只能改变 module file location，不改变断言、fixtures 或 production code。
-3. `src/core/providers/azure/batches/mod.rs` 和 `src/core/providers/azure/batches/batches_tests.rs` 必须低于 800 行。
-4. `cargo test core::providers::azure::batches --lib --all-features` 必须通过。
+3. `src/core/providers/jina/mod.rs` 和 `src/core/providers/jina/tests.rs` 必须低于 800 行。
+4. `cargo test core::providers::jina --lib --all-features` 必须通过。
 
 ## 验收标准
 
-- [ ] `src/core/providers/azure/batches/mod.rs` 使用 `#[path = "batches_tests.rs"] mod tests;` 引入测试。
-- [ ] `src/core/providers/azure/batches/batches_tests.rs` 包含原 inline test module body，且低于 800 行。
-- [ ] Focused Azure Batch tests 通过。
+- [ ] `src/core/providers/jina/mod.rs` 使用 `#[path = "tests.rs"] mod tests;` 引入测试。
+- [ ] `src/core/providers/jina/tests.rs` 包含原 inline test module body，且低于 800 行。
+- [ ] Focused Jina tests 通过。
 - [ ] `cargo fmt --all -- --check` 和 `cargo check --all-features --locked` 通过。
 - [ ] PR body 明确该 PR 是 #727 的 next tranche，不自动关闭 tracker issue。
 
 ## 发布说明
 
-No runtime behavior change. This is an Azure Batch test layout maintenance split for U-16 compliance.
+No runtime behavior change. This is a Jina provider test layout maintenance split for U-16 compliance.
